finalize version printing utils
